requirements.txt
absl-py==0.9.0
astor==0.8.1
cffi==1.14.0
click==7.1.1
Flask==1.1.1
gast==0.3.3
gevent==1.4.0
greenlet==0.4.15
grpcio==1.27.2
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
mock==4.0.2
numpy==1.18.2
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
pycparser==2.20
six==1.14.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==1.0.0
